Add check on private key creation.

+++ tests/tests.sh
@@ -71,6 +71,8 @@ cd "${TEMP_DIR}"
 running_log cookiecutter
 cookiecutterme
 
+test -f "${TEMP_DIR}/provision/id_rsa_my_project" || exit 1
+
 ( cd provision && ln -s ansible_deploy_variables ansible_vagrant_variables && cat ansible_vagrant_variables )
 
 running_log provisioning
